refactor(api): backend asíncrono, DTO validados y tipos de dato al día
Cierra la tanda de deuda técnica del 6 de septiembre:

- D-10: async/await de Conexion a los controllers, sin .Result ni .Wait()
- D-11: Data Annotations en los DTO y guard de cuerpo nulo; el 400 conserva
  el mismo {cod, mensaje} del resto de la API
- D-12: la regla de rol del ajuste sale de una posición del array a
  PoliticaMovimiento, con seis pruebas
- D-06 / D-07: nPrecio pasa a DECIMAL(10,2) y nTelefono a sTelefono
  VARCHAR(20), en esquema, procedimientos, seed y C#
- D-08: fuera las clases marcador vacías y EntRequestUsuario
- C-19: CategoriaData deja de devolver ex.Message al cliente
- C-20: UsuarioData y ZonaData cierran la conexión en finally

Ver 10-decisiones.md, D-40 a D-42.

+++ sisgapo-api/Entity/PanelEntity.cs
@@ -10,7 +10,7 @@ namespace Entity
         public int nProductos { get; set; }
         public int nCategorias { get; set; }
         public int nZonas { get; set; }
-        public long nValorInventario { get; set; }
+        public decimal nValorInventario { get; set; }
         public long nUnidades { get; set; }
         public int nPorVencer30 { get; set; }
         public int nVencidos { get; set; }
@@ -24,7 +24,7 @@ namespace Entity
         public string sNombreZona { get; set; }
         public int nProductos { get; set; }
         public long nUnidades { get; set; }
-        public long nValor { get; set; }
+        public decimal nValor { get; set; }
     }
 
     //Opción 03 — existencias por categoría
@@ -34,7 +34,7 @@ namespace Entity
         public string sNombreCategoria { get; set; }
         public int nProductos { get; set; }
         public long nUnidades { get; set; }
-        public long nValor { get; set; }
+        public decimal nValor { get; set; }
     }
 
     //Opción 04 — productos próximos a vencer
